feat(user-feedback): Render intro text directly after the description
